% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observation.R
\name{apps_time.plot}
\alias{apps_time.plot}
\title{apps_time.plot}
\usage{
apps_time.plot(variable, freq, data, xlab = NULL, ylab = NULL)
}
\arguments{
\item{variable}{the name of variable/ column in data that is assigned as the x coordinate of points in the time plot}

\item{data}{data frame where the variable values come from}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{frep}{the number of observations per unit of time.}
}
\value{
a line plot of time vs the assigned variable values
}
\description{
This function loads a plot of time series of the assigned variable with the time variable in (mdy format) in the data
}
\examples{
data(apps)
apps_time.plot(variable= Rating, freq = 12, data = apps)
}
