% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.CV.categorie.R
\name{as.CV.categorie}
\alias{as.CV.categorie}
\title{as.CV.categorie}
\usage{
as.CV.categorie(dataTxt, pays)
}
\arguments{
\item{dataTxt}{un tableau de donnee facon type: level,chiffre}

\item{pays}{un chaine de caracters, pour decider la langue.  "France","USA"}
}
\description{
as.CV.categorie prend un tableau de valeurs venant de Shiny et les separent pour mieux les afficher sous RMarkdown
}
