% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_project10Functions.r
\name{beta_initial}
\alias{beta_initial}
\title{Initial Beta Function}
\usage{
beta_initial(X, y)
}
\arguments{
\item{X}{The design \code{matrix} of independent variables}

\item{y}{The \code{vector} of dependent variables (Response variables)}
}
\value{
Initial estimate of the coefficient \code{vector} beta_i
}
\description{
Calculates the initial coefficient vector beta_i using the least squares formula
}
\examples{
# Example usage:
X <- matrix(c(1, 1, 1, 1, 2, 3, 4, 5), ncol = 2)
y <- c(2, 3, 4, 5)
beta_initial(X, y)
}
\author{
Destanie Pitt, Hayden Billmann, Ashton Wise
}
