% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_project10Functions.r
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(X, y, B = 20, alpha = 0.05)
}
\arguments{
\item{X}{The design \code{matrix} of independent variables}

\item{y}{The \code{vector} of dependent variables (Response variables)}

\item{B}{The number of bootstraps. Default is 20}

\item{alpha}{The significance level to obtain for the 1 - alpha confidence intervals for Beta. Default is .05}
}
\value{
The confidence interval
}
\description{
Calculates bootstrap confidence intervals for the logistic regression model coefficients.
}
\examples{
# Example usage:
X <- matrix(c(1, 1, 1, 1, 2, 3, 4, 5), ncol = 2)
y <- c(2, 3, 4, 5)
print(bootstrap(X, y, 20, .05))
}
\author{
Destanie Pitt, Hayden Billmann, Ashton Wise
}
