% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalProject11.R
\name{bootstrapCI}
\alias{bootstrapCI}
\title{Bootstrap Confidence Intervals for Regression Coefficients"}
\usage{
bootstrapCI(X, y, alpha = 0.05, B = 20)
}
\arguments{
\item{X}{A \code{matrix} or \code{data.frame} of predictors. Each column represents
a predictor, and rows correspond to observations.}

\item{y}{A \code{vector} of response values corresponding to the observations in \code{X}.}

\item{alpha}{A \code{numeric} value indicating the significance level for the confidence
intervals. The default alpha is 0.05.}

\item{B}{A \code{numeric} indicating the number of bootstrap resamples to
perform. The default number of bootstrap resamples is 20.}
}
\value{
A \code{matrix} containing the bootstrap confidence intervals for each regression
coefficient. Each row corresponds to a coefficient (including the intercept), and the
columns represent the lower and upper bounds of the confidence interval.
\describe{
\item{Lower Bound}{The lower bound of the confidence interval for each coefficient.}
\item{Upper Bound}{The upper bound of the confidence interval for each coefficient.}
}
}
\description{
Computes bootstrap confidence intervals for regression coefficients
using resampling with replacement.
}
