% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{bootStrapLogisticRegression}
\alias{bootStrapLogisticRegression}
\title{Preform bootstrap for dataset provided}
\usage{
bootStrapLogisticRegression(B = 20, alpha, y, X)
}
\arguments{
\item{B}{A \code{int} Number of bootstraps}

\item{alpha}{A \code{double} Alpha for bootstrap}

\item{y}{A \code{double} value of the 1 column matrix containing the response of interest.}

\item{X}{An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors.}
}
\description{
The user has the ability to choose (i) the significance level α to obtain for
the 1−α confidence intervals for β, and (ii) the number of bootstraps which
by default will be 20.
}
