% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summary_statistics2.R
\name{calculate_summary_statistics2}
\alias{calculate_summary_statistics2}
\title{Calculate summary statistics}
\usage{
calculate_summary_statistics2(climate_change, numeric_vars = NULL)
}
\arguments{
\item{climate_change}{putting your data}

\item{numeric_vars}{it allows you to provide numerical variables}
}
\value{
List a summary table
}
\description{
It calculates summary statistics and it returns a summary table
}
\examples{
calculate_summary_statistics2(climate_change)



}
