% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cat}
\alias{cat}
\title{Donnees pour les differents categories de Fanny}
\format{
Un dataframe avec 5 lines et 2 variables:
\describe{
\item{France}{Tout les categories en francais}
\item{USA}{Tout les categories en anglais}
}
}
\usage{
cat
}
\description{
Un dataframe qui contient les categories qui peut etre choisi
}
\keyword{datasets}
