% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catMailComplet.R
\name{catMailComplet}
\alias{catMailComplet}
\title{catMailComplet}
\usage{
catMailComplet(mail, langue)
}
\arguments{
\item{mail}{mail doit etre un string de caracters avec une arobase}

\item{Langue}{Langue doit etre un string de caracters}
}
\description{
catMailComplet prend les valeurs et test pour voir si il y a une arobase.
Si il n'y a pas de arobase, il renvoit une message d'erreur
}
\examples{
catNomComplet("McKenna.kevin.gmail.com","France")
[1] Erreur: pas une format propre pour un mail

}
