% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catNomComplet.R
\name{catNomComplet}
\alias{catNomComplet}
\title{catNomComplet}
\usage{
catNomComplet(Langue, nom, prenom)
}
\arguments{
\item{Langue}{Langue doit etre un string de caracters}

\item{nom}{nom doit etre un string de caracters}

\item{prenom}{prenom doit etre un string de caracters}
}
\description{
catNomComplet prend les valeurs du nom et prenom et les concatent ensembles.
Si la langue donne est France, il met le nom en majuscule
}
\examples{
catNomComplet("France","McKenna","Kevin")
[1] "MCKENNA Kevin"
catNomComplet("USA","McKenna","Kevin")
[1] "McKenna Kevin"

}
