% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catNumeroComplet.R
\name{catNumeroComplet}
\alias{catNumeroComplet}
\title{catNumeroComplet}
\usage{
catNumeroComplet(countryCode, tonNumero, Langue)
}
\arguments{
\item{countryCode}{nom doit etre un chaine de caracters de chiffres}

\item{tonNumero}{tonNumero doit etre un chaine de caracters de 10 chiffres}

\item{Langue}{Langue doit etre un string de caracters}
}
\description{
catNomComplet prend les valeurs du countryCode et tonNumero et les concatent ensembles.
Il le met dans la format du pays choisi (soit France, soit USA)
}
\examples{
catNumeroComplet("33","0123456789","France")
[1] "+33 1-23-45-67-89"
catNumeroComplet("33","0123456789","USA")
[1] "+33 012-345-6789"

}
