% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_code.R
\docType{data}
\name{class_packages}
\alias{class_packages}
\title{Load packages for a specific SDS course}
\format{
An object of class \code{list} of length 4.
}
\usage{
class_packages
}
\arguments{
\item{sds_class}{A string, like "sds_291","sds_100","sds_220","sds_192"}

\item{install_missing}{If TRUE, install any packages that aren't currently available}
}
\value{
A named logical vector showing which packages were successfully loaded
}
\description{
Given a course name like "sds_291", this will load (and optionally install) the packages needed for that course.
}
\keyword{datasets}
