% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_ci.R
\name{coef_ci}
\alias{coef_ci}
\title{Confidence Intervals of Coefficients}
\usage{
coef_ci(data, group_number, bootstrap_times)
}
\arguments{
\item{data}{data.frame}

\item{group_number}{number of groups to be devided}

\item{bootstrap_times}{number of bootstrap}
}
\value{
Confidence Intervals
}
\description{
Confidence Intervals of Coefficients
}
\examples{
library(nycflights13)
x = flights$dep_delay
y = flights$arr_time
data = data.frame(x,y)
group_number = 10
bootstrap_times = 10
coef_ci(data, group_number, bootstrap_times)
}
