% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaisonChiffres.R
\name{comparaisonChiffres}
\alias{comparaisonChiffres}
\title{comparaisonChiffres}
\usage{
comparaisonChiffres(numero)
}
\arguments{
\item{numero}{un string que est cense etre que les chiffres}
}
\description{
comparaisonChiffres
}
