% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_project10Functions.r
\name{compute_metrics}
\alias{compute_metrics}
\title{Metric Computation}
\usage{
compute_metrics(X, beta_hat, y)
}
\arguments{
\item{X}{The design \code{matrix} of independent variables}

\item{beta_hat}{The \code{vector} of estimated coefficients}

\item{y}{The \code{vector} of dependent variables (Response variables)}
}
\value{
The key performance metrics for the logistic regression model as a \code{list}
\describe{
     \item{prevalence}{The proportion of positive cases in the response}
     \item{confusion_matrix}{Representation of predicted vs actual classes}
     \item{accuracy}{The proportion of correctly classified instances out of the total instances}
     \item{sensitivity}{The proportion of actual positives correctly identified by the model}
     \item{specificity}{Measures the proportion of true negatives identified by the model}
     \item{false_discovery_rate}{The proportion of predicted positives that are actually negative}
     \item{diagnostic_odds_ratio}{Single metric that sumarizes the performance of the diagnostic test}
}
}
\description{
Calculates the confusion matrix using the predictors and response variables, then
calculates the key performance metrics for the logistic regression model
}
\examples{
# Example usage:
X <- matrix(c(1, 1, 1, 1, 2, 3, 4, 5), ncol = 2)
y <- c(2, 3, 4, 5)
b <- beta_initial(X,y)
print(compute_metrics(X,b,y))
}
\author{
Destanie Pitt, Hayden Billmann, Ashton Wise
}
