% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country}
\alias{country}
\title{Donnes pour les differents pays}
\format{
Un vector avec 1 lignes et 2 variables
\describe{
\item{France}{l'option de France}
\item{USA}{l'option de USA}
}
}
\usage{
country
}
\description{
Un vector qui contient les options des differents pays a choisir
}
\keyword{datasets}
