\name{create_account}
\alias{create_account}
\title{Create a new bank account}
\description{Create a new bank account}
\usage{
create_account(name, balance)
}
\arguments{
  \item{name}{The account name}
  \item{balance}{Initial balance with a default of 0}
}
\examples{
creat_account(checking, 400)
}
