% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_correlation_heatmap.R
\name{create_correlation_heatmap}
\alias{create_correlation_heatmap}
\title{Correlation heatmap}
\usage{
create_correlation_heatmap(climate_change, numeric_vars = NULL)
}
\arguments{
\item{climate_change}{putting your data}

\item{numeric_vars}{it allows you to provide numerical variables}
}
\value{
a heatmap
}
\description{
It creates a correlation heatmap
}
\examples{
create_correlation_heatmap(climate_change)


}
