% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherof10days.R
\name{dayweather}
\alias{dayweather}
\title{Question 4: The weather for only 10 days}
\usage{
dayweather(initial, days)
}
\arguments{
\item{initial}{numeric positive integer}

\item{days}{numeric positive integer}
}
\description{
Specifically, given an initial day’s weather conditions
Simulate the following 10 days of weather and calculate the projected rainfall accumulation in inches
}
\examples{
dayweather(1,10)

}
\author{
Edem Defor
}
