% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discord_DailyTop.R
\name{Discord_DailyTop}
\alias{Discord_DailyTop}
\title{Discord_DailyTop}
\usage{
Discord_DailyTop(df, datetime_col = "Timestamp_clean", user_col = "Author")
}
\arguments{
\item{df}{Message History Dataframe}

\item{datetime_col}{Datetime Col (Should be cleaned to proper format)}

\item{user_col}{Message Authors ("Author" by default)}
}
\description{
Generates bar chart showing how many messages the top user sent each day.
}
\examples{
Discord_DailyTop(df)

}
