% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discord_HeatMap.R
\name{Discord_Heatmap}
\alias{Discord_Heatmap}
\alias{Discord_HeatMap}
\title{Discord_HeatMap}
\usage{
Discord_HeatMap(df, datetime_col = "Timestamp_clean")
}
\arguments{
\item{df}{Message Dataframe}

\item{datetime_col}{Datetime Col (Should be cleaned to proper format)}
}
\description{
Generates a Heat Map of the Day x Hour messaging activity
}
\examples{
Discord_HeatMap(df)

}
