% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discord_HourDensity.R
\name{Discord_HourDensity}
\alias{Discord_HourDensity}
\title{Discord_HourDensity}
\usage{
Discord_HourDensity(df, datetime_col = "Timestamp_clean")
}
\arguments{
\item{df}{Message History Dataframe}

\item{datetime_col}{Datetime Column (Should be cleaned to proper format)}
}
\description{
Generates a density plot to show which times of day messages are most frequently sent.
}
\examples{
Discord_HourDensity(df)

}
