% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discord_Record.R
\name{Discord_Record}
\alias{Discord_Record}
\title{Discord_Record}
\usage{
Discord_Record(
  df,
  datetime_col = "Timestamp_clean",
  user_col = "Author",
  search_window = 36
)
}
\arguments{
\item{df}{Message History Dataframe}

\item{user_col}{Author Column, "Author" by default}

\item{search_window}{Time period to search (In Hours)}

\item{time_col}{Datetime Column, "Timestamp_Clean" by default}
}
\description{
Identifies the record amount of messages sent within the specified time period by each user, and ranks the users with the highest record.
}
\examples{
Discord_Record(df)
}
