% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discord_TimeConvert.R
\name{Discord_TimeConvert}
\alias{Discord_TimeConvert}
\title{Discord_TimeConvert}
\usage{
Discord_TimeConvert(df, column = "Date")
}
\arguments{
\item{df}{Dataframe to be cleaned}

\item{column}{Date Column to be chosen. "Date" by default.}
}
\description{
Converts the Generated Timestamps from ISO 8601 structure with fractional seconds into proper POSIXct datetimes. Suitable for generating Time Series Analysis
}
\examples{
Discord_TimeConvert(df, "Date")
}
