% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discord_TopDays.R
\name{Discord_TopDays}
\alias{Discord_TopDays}
\title{Discord_TopDays}
\usage{
Discord_TopDays(df, datetime_col = "Timestamp_clean")
}
\arguments{
\item{df}{Message History Dataframe}

\item{datetime_col}{Datetime Column (Should be cleaned to proper format)}
}
\description{
Generates a time series showing activity levels each day.
}
\examples{
Discord_TopDays(df)
}
