% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discord_TopUsers.R
\name{Discord_TopUsers}
\alias{Discord_TopUsers}
\title{Discord_TopUsers}
\usage{
Discord_TopUsers(df, user_col = "Author", time_col = "Timestamp_clean", n = 10)
}
\arguments{
\item{df}{Message History Dataframe}

\item{user_col}{Author Column, "Author" by default}

\item{time_col}{Datetime Column, "Timestamp_Clean" by default}

\item{n}{Number of users to sample}
}
\value{
Bar plot of users ranked
}
\description{
Creates ggplot visualization of top messagers. Can control column names, and amount of users that are shown
}
\details{
Discord Visualization Function
}
\examples{
Discord_TopUsers(df)
}
