% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discord_TopUserActivity.R
\name{Discord_TopUsersActivity}
\alias{Discord_TopUsersActivity}
\alias{Discord_TopUserActivity}
\title{Discord_TopUsersActivity}
\usage{
Discord_TopUserActivity(
  df,
  datetime_col = "Timestamp_clean",
  user_col = "Author",
  top_n = 5,
  smooth = FALSE
)
}
\arguments{
\item{df}{Message History Dataframe}

\item{datetime_col}{Datetime Column (Should be cleaned to proper format)}

\item{user_col}{Message Authors ("Author" by default)}

\item{top_n}{Number of users to include}

\item{smooth}{Boolean, whether to add geom_smooth}
}
\description{
Identifies the most active users and tracks the number of messages per day throught the time period using a Time Series.
}
\examples{
Discord_TopUsersActivity(df)

}
