% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discord_TopUsersCumulative.R
\name{Discord_TopUsersCumulative}
\alias{Discord_TopUsersCumulative}
\title{Discord_TopUsersCumulative}
\usage{
Discord_TopUsersCumulative(
  df,
  datetime_col = "Timestamp_clean",
  user_col = "Author",
  top_n = 10,
  points = TRUE
)
}
\arguments{
\item{df}{Message History Dataframe}

\item{datetime_col}{Datetime Col (Should be cleaned to proper format)}

\item{user_col}{Message Authors ("Author" by default)}

\item{top_n}{Number of users to include}

\item{points}{Boolean, whether points will display on the chart}
}
\description{
Generates a time series plot visualization of the top users, showing how their overall message account increases over time.
}
\examples{
Discord_TopUsersCumulative(df)

}
