% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_year.R
\name{divide_year}
\alias{divide_year}
\title{Divide by year}
\usage{
divide_year(start, end, inc, df)
}
\arguments{
\item{start}{starting year}

\item{end}{ending year}

\item{inc}{time period length}

\item{df}{dataframe with 2 columns, one named Year}
}
\value{
List of vectors of temperatures for each time period
}
\description{
For a start, end, and interval length, return a list of vectors with temperature for each time period
}
\examples{
df<-climate_change[,c("Year", "Temp")]
divide_year(1984,2008,4,df)
}
