% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedcapReport.R
\name{downloadRedcapReport}
\alias{downloadRedcapReport}
\title{Download Redcap Report}
\usage{
downloadRedcapReport(redcapTokenName, redcapUrl, redcapReportId)
}
\arguments{
\item{redcapTokenName}{The name of the environment variable storing the Redcap API token.}

\item{redcapUrl}{The URL of the Redcap API.}

\item{redcapReportId}{The ID of the report you want to retrieve.}
}
\value{
A tibble containing the contents of the Redcap report.
}
\description{
This function downloads a Redcap report using the Redcap API.
}
\examples{
\dontrun{
# Replace 'redcapTokenName', 'redcapUrl', and 'redcapReportId' with actual values
redcap_report <- downloadRedcapReport("redcapToken", "https://redcap.emory.edu/api/", "46524")
print(redcap_report)
}

}
