% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edu}
\alias{edu}
\title{Donnees pour l'education de Billy}
\format{
Un dataframe avec 3 lignes et 2 variables
\describe{
\item{France}{l'option de France}
\item{USA}{l'option de USA}
}
}
\usage{
edu
}
\description{
Un dataframe avec les donnes pour les labels des sous requetes de Education
}
\keyword{datasets}
