\name{Engine Torque from HP}
\alias{Engine Torque from HP}
\title{Engine Torque from HP}
\usage{
enginetq.hp(hp, rpm)
}
\arguments{
\item{hp}{Horsepower in US standard units}

\item{rpm}{number of time the crankshaft rotates per minute}
}
\description{
function calculates engine torque in ft.lbs from HP
}
