% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_project.R
\name{final_project}
\alias{final_project}
\title{final project
this function generates the models from knn,
MLR and random forest models, and saves the models
on rds files in models folder.}
\usage{
final_project()
}
\description{
final project
this function generates the models from knn,
MLR and random forest models, and saves the models
on rds files in models folder.
}
