% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalProject.R, R/finalproj.R
\name{finalproj}
\alias{finalproj}
\title{Prompt-Based Functions for Modeling}
\usage{
finalproj()

finalproj()
}
\value{
Prompt Menu is Launched for Choice

Prompt Menu is Launched for Choice
}
\description{
Initialize Prompt-Based Modeling (Description)

Initialize Prompt-Based Modeling (Description)
}
\details{
The user is able to request a type of model
and then input any model options and outputs that he/she
wishes.  After a dataset is entered, inputs include: variables,
interaction terms, type of distribution, and type of regression.
Outputs include: summary statistics, plots for diagnostics,
point estimates, and confidence intervals.  Basic coin flip
and dice roll functions are included for entertainment. (Details)

The user is able to request a type of model
and then input any model options and outputs that he/she
wishes.  After a dataset is entered, inputs include: variables,
interaction terms, type of distribution, and type of regression.
Outputs include: summary statistics, plots for diagnostics,
point estimates, and confidence intervals.  Basic coin flip
and dice roll functions are included for entertainment. (Details)
}
\examples{
finalproj()
finalproj()
}
\author{
c(Brian LaBarre, Silvia Jakubski)

c(Brian LaBarre, Silvia Jakubski)
}
