% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalProject-package.R
\docType{package}
\name{FinalProject-package}
\alias{FinalProject}
\alias{FinalProject-package}
\title{FinalProject: Weather Simulation}
\description{
A function to simulate the weather forecast in Richmond.
}
\author{
\strong{Maintainer}: Edem Defor \email{defore2@vcu.edu}

}
\keyword{internal}
