% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_project10Functions.r
\name{fit_model}
\alias{fit_model}
\title{Model Fit}
\usage{
fit_model(X, y)
}
\arguments{
\item{X}{The design \code{matrix} of independent variables}

\item{y}{The \code{vector} of dependent variables (Response variables)}
}
\value{
The optimized coefficients
}
\description{
Estimates the initial coefficients and uses the \code{optim()} function to minimize the
loss function and find the optimized coefficients
}
\examples{
# Example usage:
X <- matrix(c(1, 2, 3, 4, 5, 6), ncol = 3)
y <- c(0, 1)
optimized_beta <- fit_model(X, y)
print(optimized_beta)
}
\author{
Destanie Pitt, Hayden Billmann, Ashton Wise
}
