% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_project10Functions.r
\name{generate_response}
\alias{generate_response}
\title{Generate Response Variable}
\usage{
generate_response(X, beta_true)
}
\arguments{
\item{X}{A \code{matrix} of predictor variables}

\item{beta_true}{A \code{vector} of true coefficients}
}
\value{
The random binary responses for each observation
}
\description{
Calculates the predicted probabilities for each observation, then generates random
binary responses for each observation based on the probabilities
}
\examples{
# Example usage:
beta_true <- c(0.5, -1, 2)
X <- matrix(c(1, 2, 3, 4, 5, 6), ncol = 3)
response <- generate_response(X, beta_true)
print(response)
}
\author{
Destanie Pitt, Hayden Billmann, Ashton Wise
}
