% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalProject11.R
\name{get_predicted_prob}
\alias{get_predicted_prob}
\title{Predict Probabilities for Logistic Regression}
\usage{
get_predicted_prob(beta, X)
}
\arguments{
\item{beta}{A \code{vector} of regression coefficients, including the intercept.}

\item{X}{A \code{matrix} or \code{data.frame} of predictors. Each column represents
a predictor, and rows correspond to observations.}
}
\value{
A \code{vector} of predicted probabilities for each observation. The values represent
the estimated probability of the positive class (class 1).
}
\description{
Computes predicted probabilities for a logistic regression model
given a set of regression coefficients and predictors.
}
