% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goal}
\alias{goal}
\title{Donnees pour l'objectif de Billy}
\format{
Un vector avec 1 lignes et 2 variables
\describe{
\item{France}{l'option de France}
\item{USA}{l'option de USA}
}
}
\usage{
goal
}
\description{
Un dataframe avec les donnees pour les labels de l'objectif
}
\keyword{datasets}
