% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.CV.FR.R
\name{is.CV.FR}
\alias{is.CV.FR}
\title{is.CV.FR}
\usage{
is.CV.FR(obj)
}
\arguments{
\item{obj}{Un CV qui est cense etre un CV facon francais}
}
\description{
affiche "CV.FR" si obj est un CV suivant la facon Francais
}
