% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.CV.US.R
\name{is.CV.US}
\alias{is.CV.US}
\title{is.CV.US}
\usage{
is.CV.US(obj)
}
\arguments{
\item{obj}{Un CV qui est cense etre un CV facon americain}
}
\description{
affiche "CV.US" si obj est un CV suivant la facon Americain
}
