% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{linear_reg}
\alias{linear_reg}
\title{linear_reg}
\usage{
linear_reg(response, predictor, data)
}
\arguments{
\item{response}{the vector value of variable/ column in data that is response
variable in regression (numerical value)}

\item{predictor}{the vector value of variable/ column in data that is predictor
variable in regression (continuous numerical value or dummy variable)}

\item{data}{data frame where the variable values come from}
}
\value{
The result from the regression test including numerical results,
assumption plots, and plot of fitted values in the original points
}
\description{
This function is used to fit linear models. It can tested the continuous variable and dummy variable
}
\examples{
data(apps)
linear_reg(apps$Rating,apps$Reviews,apps)
}
