% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_project10Functions.r
\name{logistic_regression_pipeline}
\alias{logistic_regression_pipeline}
\title{Logistic Regression Pipeline}
\usage{
logistic_regression_pipeline(n, lambda, beta_true, B = 20, alpha = 0.05)
}
\arguments{
\item{n}{An \code{integer} specifying the number of observations}

\item{lambda}{The rate parameter for the Poisson distribution}

\item{beta_true}{A \code{vector} of true coefficients}

\item{B}{The number of bootstraps. Default is 20}

\item{alpha}{The significance level to obtain for the 1 - alpha confidence intervals for Beta. Default is .05}
}
\value{
A \code{list} of key logistic regression model findings
\describe{
     \item{beta_hat}{The optimized coefficients}
     \item{confidence_interval}{The confidence interval}
     \item{metrics}{The key performance metrics for the logistic regression model as a \code{list}}
}
}
\description{
Calculates optimized coefficients, the confidence interval,
and performance metrics for the logistic regression model. Then it returns all of these values in a list
}
\examples{
# Example usage:
# Set parameters for logistic regression pipeline
n <- 100
lambda <- 3
beta_true <- c(0.5, -1, 2)
B <- 50
alpha <- 0.05
results <- logistic_regression_pipeline(n, lambda, beta_true, B, alpha)
print(results$beta_hat)
print(results$confidence_interval)
print(results$metrics)
}
\author{
Destanie Pitt, Hayden Billmann, Ashton Wise
}
