% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikBernoulli.R
\name{logLikBernoulli}
\alias{logLikBernoulli}
\title{Compute the maximum likelihood estimate of the success probability for a Bernoulli trial}
\usage{
logLikBernoulli(data)
}
\arguments{
\item{data}{A vector of 0s and 1s representing the outcomes of Bernoulli trials.}
}
\value{
The maximum likelihood estimate of the success probability.
}
\description{
This function computes the maximum likelihood estimate of the success probability (p)
for a series of Bernoulli trials given the observed data.
}
\examples{
# Example usage:
data <- c(0, 1, 1, 0, 1)
logLikBernoulli(data)

}
