% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{message}
\alias{message}
\title{Donnees pour les messages d'erreur des Fonctions de Fanny}
\format{
Un dataframe avec 2 lignes et 2 variables:
\describe{
\item{France}{Tout les messages d'erreurs en francais}
\item{USA}{Tout les messages d'erreurs en anglais}
}
}
\usage{
message
}
\description{
Un dataframe qui contient les messages d'erreur
}
\keyword{datasets}
