% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{messageObj}
\alias{messageObj}
\title{Donnes pour le sous requete d'objectif de Billy}
\format{
Un vector avec 1 lignes et 2 variables
\describe{
\item{France}{l'option de France}
\item{USA}{l'option de USA}
}
}
\usage{
messageObj
}
\description{
Un dataframe avec les donnees pour les labels du sous requete de l'objectif
}
\keyword{datasets}
