% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Compute minimum sample size for t-tests}
\usage{
minimumN(x1, x2 = NULL)
}
\arguments{
\item{x1}{A numeric vector representing the data for one-sample or the first group for two-sample t-tests.}

\item{x2}{A numeric vector representing the data for the second group for two-sample t-tests (optional).}
}
\value{
The minimum sample size required to achieve the specified power.
}
\description{
This function computes the minimum sample size required to achieve a specified power for one-sample or two-sample t-tests.
}
\examples{
# For one-sample test
x1 <- rnorm(20)
minimumN(x1)

# For two-sample test
x2 <- rnorm(30)
minimumN(x1, x2)

}
