% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixnorm.R
\name{mixnorm}
\alias{mixnorm}
\title{Mixture normal distribution}
\usage{
mixnorm(p, mu1, mu2, s1, s2, n)
}
\arguments{
\item{p}{numeric positive integer}

\item{mu1}{numeric positive integer}

\item{mu2}{numeric positive integer}

\item{s1}{numeric positive integer}

\item{s2}{numeric positive integer}

\item{n}{numeric positive integer}
}
\description{
Generate n observations from a mixture of two normal distributionsGenerate n observations from a mixture of two normal distributions
}
\examples{
Y1 = mixnorm(0.5,-2,3,2,1.5,1000)
}
\author{
Edem Defor
}
