% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.R
\name{mlr}
\alias{mlr}
\title{Prompt for Regression Modeling}
\usage{
mlr()
}
\description{
Initialize Prompt-Based Regression Modeling (Description)
}
\details{
The user is able to request a type of model
and then input any model options and outputs that he/she
wishes.  After a dataset is entered, inputs include: variables,
interaction terms, type of distribution, and type of regression.
Outputs include: summary statistics, plots for diagnostics,
point estimates, and confidence intervals.  Basic coin flip
and dice roll functions are included for entertainment. (Details)
}
\examples{
mlr()
}
\author{
c(Brian LaBarre, Silvia Jakubski)
}
