% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_log_regression.R
\name{multi_log_regression}
\alias{multi_log_regression}
\title{Multinominal Logistic Regression}
\usage{
multi_log_regression(df)
}
\arguments{
\item{df}{: data frame
Training a Multinominal Logistic Regression model}
}
\value{
model.multinom
}
\description{
Multinominal Logistic Regression
}
