% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeroAmericain.R
\name{numeroAmericain}
\alias{numeroAmericain}
\title{numeroAmericain}
\usage{
numeroAmericain(numero)
}
\arguments{
\item{numero}{numero doit etre 10 chiffres de longeurs}
}
\description{
numeroAmericain
}
\examples{
numeroAmericain(1234567898)
}
