% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeroFrancais.R
\name{numeroFrancais}
\alias{numeroFrancais}
\title{numeroFrancais}
\usage{
numeroFrancais(numero)
}
\arguments{
\item{numero}{numero doit etre 10 chiffres de longeurs}
}
\description{
numeroFrancais
}
\examples{
numeroFrancais(1234567898)
numeroFrancais(0123456789)
}
