% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalProject11.R
\name{optimization_fn}
\alias{optimization_fn}
\title{Optimize Logistic Regression Coefficients}
\usage{
optimization_fn(X, y)
}
\arguments{
\item{X}{A \code{matrix} or \code{data.frame} of predictors. Each column represents
a predictor, and rows correspond to observations.}

\item{y}{A \code{vector} of binary response values corresponding to the observations in \code{X}.
The values should be either 0 (negative class) or 1 (positive class).}
}
\value{
A \code{vector} containing the estimated regression coefficients (including the intercept).
}
\description{
Estimates the coefficients of a logistic regression model using numerical optimization.
This function minimizes the negative log-likelihood of the logistic regression model.
}
