% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paysdf}
\alias{paysdf}
\title{Donnees pour les fonctions}
\format{
A data frame avec  17 lignes et 2 variables:
\describe{
\item{France}{Tout les question en francais}
\item{USA}{Tout les questions en anglais}
}
}
\usage{
paysdf
}
\description{
Un dataframe qui contient les valeurs pour les questions et reponse generique
dependant de la langue choisi
}
\keyword{datasets}
