% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_co2.R
\name{plot_co2}
\alias{plot_co2}
\title{Plot CO2 ovet time}
\usage{
plot_co2(data, title)
}
\arguments{
\item{data}{A data frame with CO2 data}

\item{title}{The title of the plot}
}
\value{
A ggplot2 plor
}
\description{
It creates a correlation heatmap
}
\examples{
plot_co2(climate_change,"CO2 Levels Over month")

}
