% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observation.R
\name{plot_data}
\alias{plot_data}
\title{plot_data}
\usage{
plot_data(x, y, group = NULL, data, xlab = NULL, ylab = NULL)
}
\arguments{
\item{x}{the name of variable/ column in data that is assigned as the x coordinate of points in the plot}

\item{y}{the name of variable/ column in data that is assigned as the y coordinate of points in the plot}

\item{group}{name vector of variable/column defines/groups the data (character variable)}

\item{data}{data frame where the variable values come from}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}
}
\value{
a scatter plot of 2 numeric variables corresponding the x and y coordinate
#' @examples
data(apps)
plot_data(x=Rating, y=Size, data = apps)
}
\description{
This function loads a scatter plot of 2 assigned numeric variable from the data that
is defined by another variable into group. The summary should be grouped by a character variable.
}
