% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temperature_time_series.R
\name{plot_temperature_time_series}
\alias{plot_temperature_time_series}
\title{Plot Temperature time series}
\usage{
plot_temperature_time_series(climate_change)
}
\arguments{
\item{climate_change}{the name of your data}
}
\value{
A time series plot
}
\description{
This function takes a dataset with "Year" and"Temp" columns
}
\examples{
plot_temperature_time_series(climate_change)
}
