% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plotMetric}
\alias{plotMetric}
\title{Gives ability to plot certain metrics from logistic regression Model}
\usage{
plotMetric(y, X, metric)
}
\arguments{
\item{y}{A \code{double} value of the 1 column matrix containing the response of interest.}

\item{X}{An \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors.}

\item{metric}{An \code{string} value of the metric the user would like to graph, this can be ""Prevalence", "Accuracy",
"Sensitivity", "Specificity", "False Discovery Rate", "Diagnostic Odds Ratio"}
}
\description{
This function plots the metric given as a paremeter evaluted over a grid of cut-off values
for prediction going from .1 to .9 with steps of .1
}
