% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{poly_reg}
\alias{poly_reg}
\title{poly_reg}
\usage{
poly_reg(response, predictor, order)
}
\arguments{
\item{response}{the vector value of variable/ column in data that is response
variable in regression (numerical value)}

\item{predictor}{the vector value of variable/ column in data that is predictor
variable in regression (numerical value )}

\item{dorder}{the number of ordered power assigned for the predictors}
}
\value{
The result from the regression test including numerical results, 95% confidence interval, and
assumption plots.
}
\description{
This function is used to polynomial regression models. The model is the linear combination of the predictor in ascending ordered power.
}
\examples{
data(apps)
poly_reg(apps$Rating, apps$Size, order =3)
}
