% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_forest.R
\name{rand_forest}
\alias{rand_forest}
\title{Random Forest Model}
\usage{
rand_forest(df)
}
\arguments{
\item{df}{: data frame
Training a random Forest model}
}
\value{
model.rf
}
\description{
Random Forest Model
}
