% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randforcalc.R
\name{randomforest}
\alias{randomforest}
\alias{randforcalc}
\title{correlation}
\usage{
randforcalc(df)
}
\arguments{
\item{df}{data frame}
}
\value{
plots
}
\description{
random forest analysis
}
\details{
this program takes input as data frame and performs random forest analysis on the data.
}
\author{
Anushka Jain
}
