% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcoin.R
\name{rcoin}
\alias{rcoin}
\title{Flip a Coin}
\usage{
rcoin()
}
\value{
String: Either "Heads" or "Tails"
}
\description{
Function that flips a two-sided coin (Description)
}
\details{
By typing the function, without parameters, a randomly
simulated choice between "Heads" or "Tails" is returned, 
simulating flipping a coin.
}
\examples{
rcoin()
}
\author{
c(Brian LaBarre, Silvia Jakubski)
}
