% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdie.R
\name{rdie}
\alias{rdie}
\title{Roll a Die}
\usage{
rdie()
}
\value{
integer between 1-6
}
\description{
Function that rolls a 6-sided die (Description)
}
\details{
By typing the function, without parameters, a randomly
simulated number from 1-6 is returned, simulating rolling
a 6-sided die.
}
\examples{
rdie()
}
\author{
c(Brian LaBarre, Silvia Jakubski)
}
