% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.CV.FR.R
\name{show.CV.FR}
\alias{show.CV.FR}
\title{show.CV.FR}
\usage{
\method{show}{CV.FR}(obj, option = 1)
}
\arguments{
\item{obj}{Un dataframe d'objets a afficher}

\item{option}{l'option du graphique a sortir}
}
\description{
Affiche un CV facon Francais
}
