% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.CV.US.R
\name{show.CV.US}
\alias{show.CV.US}
\title{show.CV.US}
\usage{
\method{show}{CV.US}(obj, option = 1)
}
\arguments{
\item{obj}{Un dataframe d'objets a afficher (doit etre quelque chose du classe CV.US)}

\item{option}{l'option du graphique a sortir}
}
\description{
affiche le CV facon US
}
