% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skills}
\alias{skills}
\title{Donnees pour les skills de Fanny}
\format{
Un dataframe avec 5 lines et 2 variables:
\describe{
\item{France}{Tout les questions en francais}
\item{USA}{Tout les questions en anglais}
}
}
\usage{
skills
}
\description{
Un dataframe qui contient les types de skills dependant de la langue choisi
}
\keyword{datasets}
