% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smod.R
\name{smod}
\alias{smod}
\title{Prompt for Survival Data Modeling}
\usage{
smod()
}
\description{
Initialize Prompt-Based Survival Data Modeling (Description)
}
\details{
The user is able to request a type of model
and then input any model options and outputs that he/she
wishes.  After a dataset is entered, inputs include: variables,
interaction terms, type of distribution, and type of regression.
Outputs include: summary statistics, plots for diagnostics,
point estimates, and confidence intervals.  Basic coin flip
and dice roll functions are included for entertainment. (Details)
}
\examples{
smod()
}
\author{
c(Brian LaBarre, Silvia Jakubski)
}
