% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeNames.R
\name{standardizeNames}
\alias{standardizeNames}
\title{Standardize variable names to small camel case}
\usage{
standardizeNames(data)
}
\arguments{
\item{data}{A tibble or data frame}
}
\value{
A tibble with variable names standardized to small camel case
}
\description{
Standardize variable names to small camel case
}
\examples{
data <- tibble::tibble(
  This_Is_A_Test = 1,
  Another_Test = 2
)
standardizeNames(data)
}
