% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observation.R
\name{summary_data}
\alias{summary_data}
\title{summary_data}
\usage{
summary_data(variable, group, data)
}
\arguments{
\item{variable}{name of variable/column in data that is needed to summarize (numeric variable)}

\item{group}{name vector of variable/column defines/groups the data (character variable)}

\item{data}{a data frame where the variable and group values come from}
}
\value{
A tibble table with column of group variable's elements
and the summarized variable's sample size, mean, standard deviation, min, max
}
\description{
This function loads a summary of specific variables/column in data that is
defined by another variable into group. The summary should be grouped by a character variable
}
\examples{
data(apps)
summary_data(variable = Size, group = Type, data = apps)
}
