% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survmod.R
\name{survmod}
\alias{survmod}
\title{Survival Model Choice}
\usage{
survmod(x, frmla, dtst)
}
\arguments{
\item{x}{input for choice of model}

\item{frmla}{model formula as string}

\item{dtst}{dataset for model}
}
\value{
a survival model with inputted specifications
}
\description{
Select and Compute Survival Model (Description)
}
\details{
The user inputs which model is chosen with the x parameter.
Then, the user provides the model formula and dataset.
A model is then created with those specifications (Details)
}
\examples{
survmod(1,"Surv(time,status)~1",rats)
}
\author{
c(Brian LaBarre, Silvia Jakubski)
}
