% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_plot22.R
\name{time_series_plot22}
\alias{time_series_plot22}
\title{Plot Temperature time series}
\usage{
time_series_plot22(data, variable, title)
}
\arguments{
\item{data}{A data frame with time series data}

\item{variable}{The variable to plot over time}

\item{title}{The title of the plot}
}
\value{
A ggplot2 time series plot
}
\description{
This function creates Time series Plot
}
\examples{
time_series_plot22(climate_change,CO2,"cO2 ANALYSIS")
}
