% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale a scaled vector}
\usage{
unscale(x)
}
\arguments{
\item{x}{A numeric vector to be unscaled.}
}
\value{
The unscaled vector.
}
\description{
This function unscales a vector that was previously scaled using the `scale` function in R.
}
\examples{
scaled_vector <- scale(c(1, 2, 3, 4, 5))
unscaled_vector <- unscale(scaled_vector)
unscaled_vector
# Output:  1  2  3  4  5

}
