% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weathersimulation.R
\name{weathersimulate}
\alias{weathersimulate}
\title{Weather Simulation (1000 times)}
\usage{
weathersimulate(N, initial, days)
}
\arguments{
\item{N}{numeric positive integer}

\item{initial}{numeric positive integer}

\item{days}{numeric positive integer}
}
\description{
Question 4: Repeat this 1000 times and return the average of all these simulations.
}
\examples{
weathersimulate(1000,1,10)

}
\author{
Edem Defor
}
