% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{work}
\alias{work}
\title{Donnees pour l'experience Pro de Billy}
\format{
Un vector avec 3 lignes et 2 variables
\describe{
\item{France}{l'option de France}
\item{USA}{l'option de USA}
}
}
\usage{
work
}
\description{
Un dataframe avec les donnes pour les labels des sous requetes de Experience Professionel
}
\keyword{datasets}
