% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_project10Functions.r
\name{generate_predictors}
\alias{generate_predictors}
\title{Predictor Variable Generation}
\usage{
generate_predictors(n, lambda)
}
\arguments{
\item{n}{An \code{integer} specifying the number of observations}

\item{lambda}{The rate parameter for the Poisson distribution}
}
\value{
Predictor variables \code{matrix} with 2 columns
\describe{
     \item{Column 1}{Intercept column}
     \item{Column 2}{Normally distributed predictor variables}
     \item{Column 3}{Poisson-distributed predictor variables}
}
}
\description{
This function generates a matrix of predictor variables for use in logistic regression models.
It creates a matrix with an intercept column, a column of normally distributed random variables,
and a column of Poisson-distributed random variables.
}
\examples{
# Example usage:
# Generate 5 observations with Poisson distribution parameter lambda = 3
predictors <- generate_predictors(5, 3)
print(predictors)
}
\author{
Destanie Pitt, Hayden Billmann, Ashton Wise
}
